+++ inst/app/server.R
@@ -1,31 +1,57 @@
+library(TraceR)
+
 shinyServer(function(input, output, session) {
   # Reactive graph element that is updated regularly
   graph <- reactiveVal()
+  upload_description <- reactiveVal()
 
   # Create example graph after click on button
-  observe(graph(createExampleGraph())) %>% bindEvent(input$generate_flowchart)
+  observeEvent(input$generate_flowchart,
+               graph(createExampleGraph()))
 
   # Automatically render the graph after updates
-  output$flowchart <- renderGrViz({
-    renderFlowchart(graph)
+  output$flowchart <- DiagrammeR::renderGrViz({
+    withProgress(renderFlowchart(graph),
+                 message = "Rendering graph...",
+                 value = 0.75)
   })
 
   # Download and upload
-  downloadModuleServer("download_unsigned", graph=graph)
-  uploadedGraph <- importModuleServer("import")
-  updateGraph(graph=graph, uploadedGraph=uploadedGraph)
+  downloadModuleServer("download_unsigned", graph = graph)
+
+  # export inputs and graph
+  DataTools::downloadModelServer("session_download",
+                      dat = reactive(asGraphList(graph())),
+                      inputs = input,
+                      model = reactive(NULL),
+                      rPackageName = config()[["rPackageName"]],
+                      defaultFileName = config()[["defaultFileName"]],
+                      fileExtension = config()[["fileExtension"]],
+                      modelNotes = upload_description,
+                      triggerUpdate = reactive(TRUE),
+                      onlySettings = TRUE)
+
+  uploaded_data <- importModuleServer("import")
+  updateGraph(graph = graph, uploadedGraph = reactive(uploaded_data$graph))
+  updateInput(input, output, session, uploaded_inputs = reactive(uploaded_data$input))
 
   # Display clicked node id
-  displayNodeId(input, output, outputId="clickMessage")
+  displayNodeId(input, output, outputId = "clickMessage")
 
   # Custom js to add panzoom after the svg was created
-  # Does not work currently (anymore)
   shinyjs::runjs('
-    setTimeout(function(){ // 500ms timeout as svg is not immediately available
-     var element = document.querySelector("#flowchart svg");
-     panzoom(element, {
-      bounds: true,
-      boundsPadding: 0.1
-    });
-}, 500);')
+  function initializePanzoom() {
+    var element = document.querySelector("#flowchart svg");
+    if (element && element.tagName.toLowerCase() === "svg") {
+      panzoom(element, {
+        bounds: true,
+        boundsPadding: 0.1
+      });
+    } else {
+      // Retry after a short delay if the SVG is not available yet
+      setTimeout(initializePanzoom, 100);
+    }
+  }
+  initializePanzoom();  // Initial call to check and start panzoom
+')
 })