+++ NEWS.md
@@ -1,3 +1,17 @@
+# TraceR 24.09.0
+
+## Bug Fixes
+- fix issue with zoom (#17)
+- add logo as icon for the browser tab
+
+# TraceR 24.07.0
+
+## New Features
+
+- option to import a graph from a json file from url, Pandora or file (#16)
+- option to download all user inputs together with the graph as zip file (`.tracer`) (#16)
+- option to import user inputs together with the graph from a zip file (`.tracer`) (#16)
+
 # TraceR 24.06.1
 
 ## New Features